+++ utils/config.go
@@ -1,9 +1,7 @@
 package utils
 
 import (
-	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
 )
 
@@ -12,7 +10,7 @@ var configFolder string = ".deepsource/analyzer"
 // Extracts the path of the config sent as a parameter in the user repo
 // Checks in the current working directory as well as the root directory
 // of the project
-func extractProjectRootPath() (string, error) {
+func ExtractProjectRootPath() (string, error) {
 	// Fetch the top-level directory
 	output, err := exec.Command("git", "rev-parse", "--show-toplevel").Output()
 	if err != nil {
@@ -21,19 +19,4 @@ func extractProjectRootPath() (string, error) {
 
 	// Removing trailing null characters
 	return strings.TrimRight(string(output), "\000\n"), nil
-}
-
-func InitAnalyzerConfigurationPaths() (string, string, string) {
-	cwd, _ := os.Getwd()
-
-	// Extracting the path of the project root
-	projectRoot, err := extractProjectRootPath()
-	if err != nil {
-		projectRoot = cwd
-	}
-
-	// Configuring the paths of analyzer.toml and issues directory
-	analyzerTOMLPath := filepath.Join(projectRoot, configFolder, "analyzer.toml")
-	issuesDirPath := filepath.Join(projectRoot, configFolder, "issues/")
-	return projectRoot, analyzerTOMLPath, issuesDirPath
 }
